Issue #1461 Remove redundant c8y-config-plugin code
